adding a few features
new:  previous button to move back to previous question,  skip button to skip a question, 
          continue button that allows the user to move on to next question when answered correctly
TO-DO: hint button, add links (in JSON) for users to follow to get info that will help answer the questions.
button should reveal a link for them to follow should they choose.  Maybe a way to 'close' the box it's resting in?